# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: GraphQL Schema Registry
  version: 0.1.0
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /v1/graphs:
    get:
      operationId: listGraphs
      summary: List graphs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: createGraph
      summary: Create a graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGraphRequest"
      responses:
        "201":
          description: Created
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Graph"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/graphs/{graphSlug}:
    get:
      operationId: getGraph
      summary: Get a graph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
      responses:
        "200":
          description: OK
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Graph"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    put:
      operationId: updateGraph
      summary: Update a graph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGraphRequest"
      responses:
        "200":
          description: Updated
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Graph"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      operationId: deleteGraph
      summary: Delete a graph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/graphs/{graphSlug}/supergraph.graphqls:
    get:
      operationId: getSupergraphSdl
      summary: Get the current supergraph SDL
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            Last-Modified:
              $ref: "#/components/headers/LastModified"
          content:
            text/plain:
              schema:
                type: string
        "304":
          description: Not Modified
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            ETag:
              $ref: "#/components/headers/ETag"
            Last-Modified:
              $ref: "#/components/headers/LastModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/graphs/{graphSlug}/subgraphs:
    get:
      operationId: listSubgraphs
      summary: List subgraphs
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubgraphList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: createSubgraph
      summary: Create a subgraph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubgraphRequest"
      responses:
        "201":
          description: Created
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subgraph"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/graphs/{graphSlug}/subgraphs/{subgraphSlug}:
    get:
      operationId: getSubgraph
      summary: Get a subgraph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/SubgraphSlug"
      responses:
        "200":
          description: OK
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subgraph"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    put:
      operationId: updateSubgraph
      summary: Update a subgraph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/SubgraphSlug"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubgraphRequest"
      responses:
        "200":
          description: Updated
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subgraph"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      operationId: deleteSubgraph
      summary: Delete a subgraph
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/SubgraphSlug"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/graphs/{graphSlug}/subgraphs/{subgraphSlug}/schema.graphqls:
    get:
      operationId: getSubgraphSchema
      summary: Get the current subgraph schema SDL
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/SubgraphSlug"
      responses:
        "200":
          description: OK
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      operationId: publishSubgraphSchema
      summary: Publish a subgraph schema SDL
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/SubgraphSlug"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        "204":
          description: Stored
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/graphs/{graphSlug}/subgraphs/{subgraphSlug}/validate-schema:
    post:
      operationId: validateSubgraphSchema
      summary: Validate a subgraph schema SDL
      parameters:
        - $ref: "#/components/parameters/GraphSlug"
        - $ref: "#/components/parameters/SubgraphSlug"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        "200":
          description: Valid
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - diagnostics
                properties:
                  diagnostics:
                    type: array
                    items:
                      type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Invalid schema
          content:
            application/problem+json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - type
                  - title
                  - status
                properties:
                  type:
                    type: string
                    format: uri-reference
                  title:
                    type: string
                  status:
                    type: integer
                    format: int32
                    minimum: 100
                    maximum: 599
                  diagnostics:
                    type: array
                    items:
                      type: string
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    bearerAuth:
      $ref: "./components/security-schemes/bearer-auth.yaml"
  headers:
    CacheControl:
      $ref: "./components/headers/cache-control.yaml"
    ETag:
      $ref: "./components/headers/etag.yaml"
    LastModified:
      $ref: "./components/headers/last-modified.yaml"
    Location:
      $ref: "./components/headers/location.yaml"
...
  parameters:
    GraphSlug:
      $ref: "./components/parameters/graph-slug.yaml"
    SubgraphSlug:
      $ref: "./components/parameters/subgraph-slug.yaml"
    IfMatch:
      $ref: "./components/parameters/if-match.yaml"
  schemas:
    Graph:
      $ref: "./components/schemas/graph.yaml"
    GraphList:
      $ref: "./components/schemas/graph-list.yaml"
    CreateGraphRequest:
      $ref: "./components/schemas/create-graph-request.yaml"
    UpdateGraphRequest:
      $ref: "./components/schemas/update-graph-request.yaml"
    Subgraph:
      $ref: "./components/schemas/subgraph.yaml"
    SubgraphList:
      $ref: "./components/schemas/subgraph-list.yaml"
    CreateSubgraphRequest:
      $ref: "./components/schemas/create-subgraph-request.yaml"
    UpdateSubgraphRequest:
      $ref: "./components/schemas/update-subgraph-request.yaml"
...
  responses:
    BadRequest:
      $ref: "./components/responses/bad-request.yaml"
    Unauthorized:
      $ref: "./components/responses/unauthorized.yaml"
    Forbidden:
      $ref: "./components/responses/forbidden.yaml"
    NotFound:
      $ref: "./components/responses/not-found.yaml"
    Conflict:
      $ref: "./components/responses/conflict.yaml"
    UnprocessableEntity:
      $ref: "./components/responses/unprocessable-entity.yaml"
    PreconditionFailed:
      $ref: "./components/responses/precondition-failed.yaml"
    InternalServerError:
      $ref: "./components/responses/internal-server-error.yaml"
...
    ServiceUnavailable:
      $ref: "./components/responses/service-unavailable.yaml"